+++ doc/source/conf.py
@@ -28,6 +28,7 @@ author = 'Geoffroy Lesur'
 release = '0.7'
 
 
+
 # -- General configuration ---------------------------------------------------
 
 # Add any Sphinx extension module names here, as strings. They can be