+++ src/idefix.hpp
@@ -65,6 +65,10 @@ using Layout = Kokkos::LayoutRight;
 #define UCT_CONTACT  3
 #define UCT_HLL      4
 
+// Runge-Kutta-Legendre switch
+#define RKL   YES
+
+
 // Basic configuration
 #include "definitions.hpp"
 #include "real_types.hpp"
@@ -205,7 +209,6 @@ using Layout = Kokkos::LayoutRight;
 
 
 #define NVAR    (NFLX)
-
 
 
 // File handler depends on the type of I/O we use